+++ src/natureObjects/clouds.js
